Traktor: Experimental "apply spline" command in terrain editor which applies selected brush along a shape spline.
[git-p4: depot-paths = "//depot/traktor/": change = 23852]

+++ code/Terrain/Editor/ColorBrush.cpp
@@ -19,7 +19,7 @@ ColorBrush::ColorBrush(drawing::Image* colorImage)
 {
 }
 
-uint32_t ColorBrush::begin(int32_t x, int32_t y, const State& state)
+uint32_t ColorBrush::begin(float x, float y, const State& state)
 {
 	m_radius = state.radius;
 	m_fallOff = state.falloff;
@@ -28,7 +28,7 @@ uint32_t ColorBrush::begin(int32_t x, int32_t y, const State& state)
 	return MdColor;
 }
 
-void ColorBrush::apply(int32_t x, int32_t y)
+void ColorBrush::apply(float x, float y)
 {
 	Color4f targetColor;
 	for (int32_t iy = -m_radius; iy <= m_radius; ++iy)
@@ -53,7 +53,7 @@ void ColorBrush::apply(int32_t x, int32_t y)
 	}
 }
 
-void ColorBrush::end(int32_t x, int32_t y)
+void ColorBrush::end(float x, float y)
 {
 }
