+++ code/Terrain/Editor/SplatBrush.h
@@ -30,11 +30,11 @@ class SplatBrush : public IBrush
 public:
 	SplatBrush(const resource::Proxy< hf::Heightfield >& heightfield, drawing::Image* splatImage);
 
-	virtual uint32_t begin(int32_t x, int32_t y, const State& state) override final;
+	virtual uint32_t begin(float x, float y, const State& state) override final;
 
-	virtual void apply(int32_t x, int32_t y) override final;
+	virtual void apply(float x, float y) override final;
 
-	virtual void end(int32_t x, int32_t y) override final;
+	virtual void end(float x, float y) override final;
 
 private:
 	resource::Proxy< hf::Heightfield > m_heightfield;
